july2020/solutions/day02_BinaryTreeLevelOrderTraversalII.py
# Given a binary tree, return the bottom-up level order traversal of its nodes' values. (ie, from left to right, level by level from leaf to root).

# For example:
# Given binary tree [3,9,20,null,null,15,7],
#     3
#    / \
#   9  20
#     /  \
#    15   7
# return its bottom-up level order traversal as:
# [
#   [15,7],
#   [9,20],
#   [3]
# ]

# Definition for a binary tree node.
class TreeNode:
    def __init__(self, val=0, left=None, right=None):
        self.val = val
        self.left = left
        self.right = right

def levelOrderBottom(root):
	result = []
	if not root:
		return result
	queue = []
	queue.append(root)
	while queue:
		nodes = []
		n = len(queue)
		for _ in range(n):
			node = queue.pop(0)
			nodes.append(node.val)
			if node.left:
				queue.append(node.left)
			if node.right:
				queue.append(node.right)
		result.append(nodes)
	return result[::-1]
